:hammer: First functionnal Pathfinding (in Log).
We have a first functionnal pathfinding. But there is a lot of default : First of all, the method getNextDirection doesnt work as expected because of the ennemy is already on the start position. So this function must be refactored ! Also we will have problem with movement of the ennemy. Because the pathfinding work in a graph with with 1,1 ratio.